Put activity into common
