+++ shared/teams/main/team-row.tsx
@@ -7,6 +7,7 @@ import * as Container from '../../util/container'
 import * as Chat2Gen from '../../actions/chat2-gen'
 import TeamMenu from '../team/menu-container'
 import {pluralize} from '../../util/string'
+import {Activity} from '../common'
 
 type Props = {
   firstItem: boolean
@@ -28,11 +29,10 @@ const TeamRow = (props: Props) => {
   const onChat = () =>
     dispatch(Chat2Gen.createPreviewConversation({reason: 'teamRow', teamname: teamMeta.teamname}))
 
-  const popupRoot = React.useRef(null)
-  const {popup, setShowingPopup, showingPopup} = Kb.usePopup(popupRoot, () => (
+  const {popup, popupRoot, setShowingPopup, showingPopup} = Kb.usePopup(getAttachmentRef => (
     <TeamMenu
       teamID={teamID}
-      attachTo={() => popupRoot.current}
+      attachTo={getAttachmentRef}
       onHidden={() => setShowingPopup(false)}
       visible={showingPopup}
     />
@@ -111,34 +111,7 @@ const TeamRow = (props: Props) => {
   )
 }
 
-type ActivityLevel = 'active' | 'recently' | 'extinct'
-const activityToIcon: {[key in ActivityLevel]: Kb.IconType} = {
-  active: 'iconfont-fire',
-  extinct: 'iconfont-rip',
-  recently: 'iconfont-team-leave',
-}
-const activityToLabel = {
-  active: 'Active',
-  extinct: 'Extinct',
-  recently: 'Recently active',
-}
-const Activity = ({level}: {level: ActivityLevel}) => (
-  <Kb.Box2 direction="horizontal" gap="xtiny" alignItems="center" fullWidth={Styles.isMobile}>
-    <Kb.Icon
-      type={activityToIcon[level]}
-      color={level === 'active' ? Styles.globalColors.greenDark : Styles.globalColors.black_50}
-      sizeType="Small"
-    />
-    <Kb.Text type="BodySmall" style={level === 'active' && styles.activityActive}>
-      {activityToLabel[level]}
-    </Kb.Text>
-  </Kb.Box2>
-)
-
 const styles = Styles.styleSheetCreate(() => ({
-  activityActive: {
-    color: Styles.globalColors.greenDark,
-  },
   bodyContainer: {
     paddingBottom: Styles.globalMargins.tiny,
     paddingTop: Styles.globalMargins.tiny,
